commands and proper role removal

+++ Domain/Models/ServerConfig.cs
@@ -3,7 +3,7 @@
     public class ServerConfig
     {
         public ulong ServerId { get; set; }
-        public int Duration { get; set; }
+        public long Duration { get; set; }
         public ulong Role { get; set; }
     }
 }
